requirements.txt
langchain==0.0.27
html2text==2020.1.16
playwright==1.34.0
python-dotenv==1.0.0
readabilipy==0.2.0
unstructured==0.7.2
tiktoken==0.4.0
tabulate==0.9.0